chore(starfish): Expose App Start in sidebar
Renames user-facing usage of "App Initialization" to "App Startup"
since it's more commonly used
