+++ static/app/views/starfish/utils/routeNames.tsx
@@ -6,7 +6,7 @@ export const ROUTE_NAMES = {
   'endpoint-overview': t('Endpoint Overview'),
   'span-summary': t('Span Summary'),
   'web-service': t('Web Service'),
-  initialization: t('App Initialization'),
+  'app-startup': t('App Startup'),
   pageload: t('Mobile'),
   responsiveness: t('Responsiveness'),
 };
